chore: :art: Organización de directorios dentro de /src
