+++ src/actualizar_documentos/actualizar_basic.py
@@ -1,7 +1,6 @@
-from conectar_con_mongo import conectar_con_mongo
+from mongo.conectar_con_mongo import conectar_con_mongo
 
-def insertar_documento_standard(nombre_pack,stock_pack,dimensiones_pack,contenido):
-    
+def actualizar_basic(nombre_pack,stock_pack,dimensiones_pack,contenido,documento):
     contenidos = {"contenidos":{
         contenido[0]:{
             "caracteristicas":{
@@ -22,29 +21,20 @@ def insertar_documento_standard(nombre_pack,stock_pack,dimensiones_pack,contenid
                 "Stock":contenido[12],
                 "Demanda":contenido[13]
             }
-        },
-        contenido[14]:{
-            "caracteristicas":{
-                "Precio":str(contenido[15]),
-                "Calidad":contenido[16],
-                "Material":contenido[17],
-                "Cantidad":contenido[18],
-                "Stock":contenido[19],
-                "Demanda":contenido[20]
-            }
         }}}
     
     colleccion = conectar_con_mongo()
     
-    colleccion.insert_one({"Nombre pack":nombre_pack,
-    "calidad":"Standard",
-    "precio":str(contenido[1]+contenido[8]+contenido[15]),
+    documento_actualizado ={"Nombre pack":nombre_pack,
+    "calidad":"Basic",
+    "precio":str(contenido[1] + contenido[8]),
     "stock":stock_pack,
     "dimensiones":{
         "altura":dimensiones_pack[0],
         "ancho":dimensiones_pack[1]
     },
     "contenidos":contenidos["contenidos"]
-    })
-    
-    print("Se ha introducido el documento.")
+    }
+
+    colleccion.update_one(documento, documento_actualizado)
+    print('El documento ha sido actualizado con extio')
