+++ src/actualizar_documentos/actualizar_standard.py
@@ -1,6 +1,7 @@
-from conectar_con_mongo import conectar_con_mongo
+from mongo.conectar_con_mongo import conectar_con_mongo
 
-def actualizar_basic(nombre_pack,stock_pack,dimensiones_pack,contenido,documento):
+def actualizar_standard(nombre_pack,stock_pack,dimensiones_pack,contenido,documento):
+    
     contenidos = {"contenidos":{
         contenido[0]:{
             "caracteristicas":{
@@ -21,13 +22,23 @@ def actualizar_basic(nombre_pack,stock_pack,dimensiones_pack,contenido,documento
                 "Stock":contenido[12],
                 "Demanda":contenido[13]
             }
+        },
+        contenido[14]:{
+            "caracteristicas":{
+                "Precio":str(contenido[15]),
+                "Calidad":contenido[16],
+                "Material":contenido[17],
+                "Cantidad":contenido[18],
+                "Stock":contenido[19],
+                "Demanda":contenido[20]
+            }
         }}}
     
     colleccion = conectar_con_mongo()
     
     documento_actualizado ={"Nombre pack":nombre_pack,
-    "calidad":"Basic",
-    "precio":str(contenido[1] + contenido[8]),
+    "calidad":"Standard",
+    "precio":str(contenido[1]+contenido[8]+contenido[15]),
     "stock":stock_pack,
     "dimensiones":{
         "altura":dimensiones_pack[0],
@@ -37,4 +48,4 @@ def actualizar_basic(nombre_pack,stock_pack,dimensiones_pack,contenido,documento
     }
 
     colleccion.update_one(documento, documento_actualizado)
-    print('El documento ha sido actualizado con extio')
+    print('El documento ha sido actualizado con exito')
