+++ src/eliminar_documentos/query_borrar_documentos.py
