+++ src/eliminar_documentos/sanear_query_borrar.py
