+++ src/encontrar_documentos/query.py
