+++ src/insertar_documentos/insertar_documento_premium.py
@@ -1,6 +1,6 @@
-from conectar_con_mongo import conectar_con_mongo
+from mongo.conectar_con_mongo import conectar_con_mongo
 
-def actualizar_standard(nombre_pack,stock_pack,dimensiones_pack,contenido,documento):
+def insertar_documento_premium(nombre_pack,stock_pack,dimensiones_pack,contenido):
     
     contenidos = {"contenidos":{
         contenido[0]:{
@@ -32,20 +32,29 @@ def actualizar_standard(nombre_pack,stock_pack,dimensiones_pack,contenido,docume
                 "Stock":contenido[19],
                 "Demanda":contenido[20]
             }
+        },
+        contenido[21]:{
+            "caracteristicas":{
+                "Precio":str(contenido[22]),
+                "Calidad":contenido[23],
+                "Material":contenido[24],
+                "Cantidad":contenido[25],
+                "Stock":contenido[26],
+                "Demanda":contenido[27]
+            }
         }}}
     
     colleccion = conectar_con_mongo()
     
-    documento_actualizado ={"Nombre pack":nombre_pack,
-    "calidad":"Standard",
-    "precio":str(contenido[1]+contenido[8]+contenido[15]),
+    colleccion.insert_one({"Nombre pack":nombre_pack,
+    "calidad":"Premium",
+    "precio":str(contenido[1]+contenido[8]+contenido[15]+contenido[22]),
     "stock":stock_pack,
     "dimensiones":{
         "altura":dimensiones_pack[0],
         "ancho":dimensiones_pack[1]
     },
     "contenidos":contenidos["contenidos"]
-    }
-
-    colleccion.update_one(documento, documento_actualizado)
-    print('El documento ha sido actualizado con exito')
+    })
+    
+    print("Se ha introducido el documento.")
