+++ src/insertar_documentos/insertar_documento_standard.py
@@ -1,6 +1,7 @@
-from conectar_con_mongo import conectar_con_mongo
+from mongo.conectar_con_mongo import conectar_con_mongo
 
-def actualizar_premium(nombre_pack,stock_pack,dimensiones_pack,contenido,documento):
+def insertar_documento_standard(nombre_pack,stock_pack,dimensiones_pack,contenido):
+    
     contenidos = {"contenidos":{
         contenido[0]:{
             "caracteristicas":{
@@ -31,29 +32,19 @@ def actualizar_premium(nombre_pack,stock_pack,dimensiones_pack,contenido,documen
                 "Stock":contenido[19],
                 "Demanda":contenido[20]
             }
-        },
-        contenido[21]:{
-            "caracteristicas":{
-                "Precio":str(contenido[22]),
-                "Calidad":contenido[23],
-                "Material":contenido[24],
-                "Cantidad":contenido[25],
-                "Stock":contenido[26],
-                "Demanda":contenido[27]
-            }
         }}}
     
     colleccion = conectar_con_mongo()
     
-    documento_actualizado = {"Nombre pack":nombre_pack,
-    "calidad":"Premium",
-    "precio":str(contenido[1]+contenido[8]+contenido[15]+contenido[22]),
+    colleccion.insert_one({"Nombre pack":nombre_pack,
+    "calidad":"Standard",
+    "precio":str(contenido[1]+contenido[8]+contenido[15]),
     "stock":stock_pack,
     "dimensiones":{
         "altura":dimensiones_pack[0],
         "ancho":dimensiones_pack[1]
     },
     "contenidos":contenidos["contenidos"]
-    }
-
-    colleccion.update_one(documento, documento_actualizado)
+    })
+    
+    print("Se ha introducido el documento.")
